refactor(core): Phase 3 — move collector & history into hwviz_core, drop AppHandle
Completes #1405:

- Add `hwviz_core::collector` module with three submodules:
  - `history::HistoryStore` — owns the CPU / memory / GPU / process
    history rings behind `Arc<Mutex<...>>`, exposes a Core read API
    (`cpu_history`, `memory_history`, `gpu_history`, `process_list`,
    `current_cpu_usage_overall`, `current_cpu_usage_per_processor`,
    `current_memory_usage_percent`) used by the App-side Tauri command
    handlers, plus a transitional `arc_handles()` accessor that hands
    the underlying `Arc`s to the (still-Tauri-resident) archive worker.
  - `sampling` — `sample_system` / `sample_gpu` / `build_metrics_snapshot`,
    relocated from `services/monitoring_service.rs`. Writes into
    `HistoryStore` instead of `MonitorResources`. Temperature unit
    conversion is removed from this layer; collectors always publish
    raw °C in `MetricsSnapshot`.
  - `system_monitor::SystemMonitorController` — relocated from
    `workers/system_monitor.rs`. `setup` signature is now
    `(Arc<HistoryStore>, EventBus, tokio::runtime::Handle)` —
    **`AppHandle` is dropped**. The runtime handle is passed in so the
    loop can spawn onto Tauri's existing tokio runtime; the App-side
    caller hands over `tauri::async_runtime::handle().inner().clone()`.

- Apply temperature unit conversion in `WindowAdapter`. The collector
  publishes °C; the App-side adapter reads `settings::AppState`'s
  `temperature_unit` per snapshot and converts to F before emitting
  `HardwareMonitorUpdate`. Adds tests for °C passthrough rounding and
  °C → °F conversion.

- Migrate `commands/hardware.rs` to take
  `tauri::State<'_, Arc<HistoryStore>>` and call the Core read API.
  Convert the returned `core::models::ProcessInfo` to the wire
  `ProcessInfo` via the `From` impl added in the previous commit.

- Construct `Arc<HistoryStore>` in `lib.rs::run` and `app.manage(...)`
  it. The archive worker keeps its existing `MonitorResources`
  contract for now — Phase 4 will replace this with an EventBus
  subscription. Drop the unused `system` field from `MonitorResources`.

- Delete the now-empty / superseded modules:
  - `src-tauri/src/services/{cpu_service, monitoring_service, process_service}.rs`
  - `src-tauri/src/workers/system_monitor.rs`
  - `src-tauri/src/models/hardware::HardwareMonitorState`
  - `services/memory_service::memory_usage_percent` (replaced by
    `HistoryStore::current_memory_usage_percent`).
  - `HARDWARE_HISTORY_BUFFER_SIZE` and
    `MAX_HISTORY_QUERY_DURATION_SECONDS` from
    `src-tauri/src/constants.rs` (now in `core::collector::history`).
  - `WorkersState::monitor` is re-typed to
    `Mutex<Option<hwviz_core::collector::SystemMonitorController>>`.

- Update `.github/instructions/rust.instructions.md` to reflect the
  workspace split (no `tauri` dep in `core`; no `specta`/`tauri_specta`
  in `core`; `log_internal` no longer needs to be imported alongside
  the level macros).

DoD verification (#1405):
- `commands/hardware.rs` no longer locks `Arc<Mutex<History*>>` — every
  history read goes through `HistoryStore` methods.
- `SystemMonitorController::setup` no longer takes `AppHandle`.
- `cargo test -p hwviz-core` passes 93 tests including the new
  `collector::history::tests::*` and `collector::sampling::tests::*`,
  with no Tauri runtime.
- `cargo tauri-fmt`, `cargo tauri-lint`,
  `cargo test --workspace --lib --tests`, `npm run lint`, `npm test`
  all pass; the dashboard, history graphs, process list, and °C / °F
  toggle behave unchanged at runtime.
