+++ src-tauri/src/services/memory_service.rs
@@ -1,21 +1,5 @@
-use crate::models::hardware::{HardwareMonitorState, MemoryInfo};
+use crate::models::hardware::MemoryInfo;
 use hwviz_core::platform::factory::PlatformFactory;
-
-///
-/// ## Return memory usage (%). Round used / total * 100
-///
-/// Returns 0 if total is 0
-///
-pub fn memory_usage_percent(state: &HardwareMonitorState) -> i32 {
-  let system = state.system.lock().unwrap();
-  let used = system.used_memory() as f64;
-  let total = system.total_memory() as f64;
-  if total == 0.0 {
-    0
-  } else {
-    ((used / total) * 100.0).round() as i32
-  }
-}
 
 ///
 /// ## Get detailed memory information via Platform
