--- src-tauri/src/services/monitoring_service.rs
@@ -1,1024 +1,0 @@
-use std::collections::VecDeque;
-use std::sync::{Arc, Mutex};
-
-use crate::constants::{
-  HARDWARE_HISTORY_BUFFER_SIZE, MAX_HISTORY_QUERY_DURATION_SECONDS,
-};
-use crate::models::hardware::HardwareMonitorState;
-use crate::models::hardware_archive::MonitorResources;
-
-/// A single GPU sample collected per physical GPU.
-/// `None` means the metric is unavailable for this GPU vendor/platform.
-pub struct GpuSample {
-  pub gpu_id: String,
-  pub name: String,
-  pub usage: Option<f32>,
-  pub temperature: Option<f32>,
-  pub dedicated_memory_kb: Option<f32>,
-  pub cooler_level: Option<u32>,
-  pub source: String,
-}
-
-pub struct SystemSample {
-  pub cpu_usage: f32,
-  pub memory_usage: f32,
-  pub processors_usage: Vec<f32>,
-}
-
-/// System sampling for one cycle (CPU/memory/process)
-pub fn sample_system(resources: &MonitorResources) -> Option<SystemSample> {
-  if let Some((cpu_usage, memory_usage, processors_usage, process_metrics)) =
-    resources.system.lock().ok().map(|mut sys| {
-      sys.refresh_all();
-
-      let cpu_usage = calculate_average_cpu_usage(sys.cpus());
-      let memory_usage =
-        calculate_memory_usage_percentage(sys.used_memory(), sys.total_memory());
-      let processors_usage: Vec<f32> = sys.cpus().iter().map(|c| c.cpu_usage()).collect();
-
-      let process_metrics: Vec<_> = sys
-        .processes()
-        .iter()
-        .map(|(pid, process)| {
-          (*pid, process.cpu_usage(), process.memory() as f32 / 1024.0)
-        })
-        .collect();
-
-      (cpu_usage, memory_usage, processors_usage, process_metrics)
-    })
-  {
-    push_history(&resources.cpu_history, cpu_usage);
-    push_history(&resources.memory_history, memory_usage);
-    update_process_histories(resources, &process_metrics);
-    Some(SystemSample {
-      cpu_usage,
-      memory_usage,
-      processors_usage,
-    })
-  } else {
-    None
-  }
-}
-
-fn calculate_average_cpu_usage(cpus: &[sysinfo::Cpu]) -> f32 {
-  match cpus.len() {
-    0 => 0.0,
-    len => (cpus.iter().map(|cpu| cpu.cpu_usage()).sum::<f32>() / len as f32).round(),
-  }
-}
-
-fn calculate_memory_usage_percentage(used: u64, total: u64) -> f32 {
-  match total {
-    0 => 0.0,
-    total => ((used as f64 / total as f64) * 100.0).round() as f32,
-  }
-}
-
-fn update_process_histories(
-  resources: &MonitorResources,
-  process_metrics: &[(sysinfo::Pid, f32, f32)],
-) {
-  let mut cpu_histories = resources.process_cpu_histories.lock().unwrap();
-  let mut mem_histories = resources.process_memory_histories.lock().unwrap();
-
-  process_metrics
-    .iter()
-    .for_each(|(pid, cpu_usage, memory_mb)| {
-      let cpu_history = cpu_histories.entry(*pid).or_default();
-      if cpu_history.len() >= HARDWARE_HISTORY_BUFFER_SIZE {
-        cpu_history.pop_front();
-      }
-      cpu_history.push_back(*cpu_usage);
-      let mem_history = mem_histories.entry(*pid).or_default();
-      if mem_history.len() >= HARDWARE_HISTORY_BUFFER_SIZE {
-        mem_history.pop_front();
-      }
-      mem_history.push_back(*memory_mb);
-    });
-}
-
-#[cfg(target_os = "windows")]
-pub async fn sample_gpu(resources: &MonitorResources) -> Vec<GpuSample> {
-  use hwviz_core::infrastructure::providers::nvapi_provider;
-  use nvapi::PhysicalGpu;
-
-  let mut gpu_metrics: Vec<GpuSample> = Vec::new();
-
-  // ── NVIDIA GPUs via NVAPI ──
-  if let Some(nvapi_metrics) = PhysicalGpu::enumerate().ok().map(|gpus| {
-    gpus
-      .iter()
-      .map(|gpu| {
-        let name = gpu.full_name().unwrap_or_else(|_| "Unknown".to_string());
-        let usage = nvapi_provider::get_gpu_usage_from_physical_gpu(gpu);
-        let temperature =
-          nvapi_provider::get_gpu_temperature_from_physical_gpu(gpu) as f32;
-        let memory_usage =
-          nvapi_provider::get_gpu_dedicated_memory_usage_from_physical_gpu(gpu) as f32;
-        let cooler_level = nvapi_provider::get_gpu_cooler_level_from_physical_gpu(gpu);
-        GpuSample {
-          gpu_id: format!("nvapi:{}", gpu.gpu_id().unwrap_or(0)),
-          name,
-          usage: Some(usage),
-          temperature: Some(temperature),
-          dedicated_memory_kb: Some(memory_usage),
-          cooler_level,
-          source: "NVAPI".to_string(),
-        }
-      })
-      .collect::<Vec<_>>()
-  }) {
-    gpu_metrics.extend(nvapi_metrics);
-  }
-
-  // ── AMD GPUs via ADL ──
-  if hwviz_core::infrastructure::providers::adl_provider::is_available() {
-    sample_amd_gpu(&mut gpu_metrics).await;
-  }
-
-  // ── Intel GPUs via PDH + DXGI ──
-  sample_intel_gpu(&mut gpu_metrics).await;
-
-  if !gpu_metrics.is_empty() {
-    update_gpu_histories(resources, &gpu_metrics);
-  }
-
-  gpu_metrics
-}
-
-/// Resolve a GPU's canonical name using a pre-built BDF→name lookup table.
-/// Returns the mapped name on hit, or the original `adl_name` on miss.
-#[cfg(any(target_os = "windows", test))]
-fn resolve_gpu_name_from_map(
-  bdf_map: &std::collections::HashMap<(i32, i32, i32), String>,
-  adl_name: &str,
-  bus: i32,
-  device: i32,
-  function: i32,
-) -> String {
-  bdf_map
-    .get(&(bus, device, function))
-    .cloned()
-    .unwrap_or_else(|| adl_name.to_string())
-}
-
-/// Build a lookup table that maps PCI BDF → DXGI device description (= the
-/// canonical GPU name used by `get_gpu_info` / `GraphicInfo`).
-///
-/// The table is computed once via SetupDi and cached for the lifetime of
-/// the process.  Because SetupDi calls are blocking Win32 APIs, the first
-/// invocation offloads them to the Tokio blocking thread pool so we never
-/// stall an async worker thread.
-#[cfg(target_os = "windows")]
-async fn bdf_to_dxgi_name() -> &'static std::collections::HashMap<(i32, i32, i32), String>
-{
-  use hwviz_core::infrastructure::providers::setupdi_provider;
-  use crate::log_error;
-
-  static MAP: tokio::sync::OnceCell<std::collections::HashMap<(i32, i32, i32), String>> =
-    tokio::sync::OnceCell::const_new();
-
-  MAP
-    .get_or_init(|| async {
-      match tokio::task::spawn_blocking(|| {
-        let adapters = setupdi_provider::enumerate_display_adapters();
-        adapters
-          .into_iter()
-          .map(|a| ((a.bus, a.device, a.function), a.description))
-          .collect()
-      })
-      .await
-      {
-        Ok(map) => map,
-        Err(e) => {
-          log_error!(
-            &format!("SetupDi enumeration task failed: {e}"),
-            "monitoring_service::bdf_to_dxgi_name",
-            None::<&str>
-          );
-          std::collections::HashMap::new()
-        }
-      }
-    })
-    .await
-}
-
-/// Collect AMD GPU usage and temperature via ADL.
-/// VRAM usage is not available via ADL.
-#[cfg(target_os = "windows")]
-async fn sample_amd_gpu(gpu_metrics: &mut Vec<GpuSample>) {
-  use hwviz_core::infrastructure::providers::adl_provider;
-
-  // Usage per adapter (with BDF)
-  let usages = adl_provider::get_amd_gpu_usage_per_adapter()
-    .await
-    .unwrap_or_default();
-
-  if usages.is_empty() {
-    // No usage metrics; skip temperature queries and SetupDi enumeration
-    return;
-  }
-  // Temperature per adapter (with BDF)
-  let temps = adl_provider::get_amd_gpu_temperatures_per_adapter()
-    .await
-    .unwrap_or_default();
-
-  // Build a temp lookup by BDF (more reliable than name matching)
-  let temp_map: std::collections::HashMap<(i32, i32, i32), f32> = temps
-    .iter()
-    .map(|m| ((m.bus, m.device, m.function), m.value))
-    .collect();
-
-  // Fetch the BDF→DXGI name map once, not per-adapter
-  let bdf_map = bdf_to_dxgi_name().await;
-
-  for metric in &usages {
-    let bdf = (metric.bus, metric.device, metric.function);
-    let temperature = temp_map.get(&bdf).copied();
-    let name = resolve_gpu_name_from_map(
-      bdf_map,
-      &metric.adapter_name,
-      metric.bus,
-      metric.device,
-      metric.function,
-    );
-    gpu_metrics.push(GpuSample {
-      gpu_id: format!("pci:{}:{}:{}", metric.bus, metric.device, metric.function),
-      name,
-      usage: Some(metric.value),
-      temperature,
-      dedicated_memory_kb: None,
-      cooler_level: None,
-      source: "ADL".to_string(),
-    });
-  }
-}
-
-/// Collect Intel GPU usage via PDH performance counters, filtered by LUID.
-#[cfg(target_os = "windows")]
-async fn sample_intel_gpu(gpu_metrics: &mut Vec<GpuSample>) {
-  use hwviz_core::infrastructure::providers::pdh_provider::{self, GpuEngineType};
-
-  let intel_gpus =
-    hwviz_core::infrastructure::providers::directx::get_intel_gpu_luid_info_cached().await;
-  if intel_gpus.is_empty() {
-    return;
-  }
-
-  for gpu in intel_gpus {
-    let usage = pdh_provider::query_gpu_usage_by_luid_and_engine(
-      gpu.luid_high,
-      gpu.luid_low,
-      GpuEngineType::Graphics3D,
-    )
-    .await
-    .ok()
-    .map(|v| (v * 100.0).round());
-
-    gpu_metrics.push(GpuSample {
-      gpu_id: format!("pdh:{}", gpu.name),
-      name: gpu.name.clone(),
-      usage,
-      temperature: None,
-      dedicated_memory_kb: None,
-      cooler_level: None,
-      source: "PDH".to_string(),
-    });
-  }
-}
-
-#[cfg(target_os = "linux")]
-pub async fn sample_gpu(resources: &MonitorResources) -> Vec<GpuSample> {
-  let gpu_metrics = collect_linux_gpu_metrics().await;
-
-  if !gpu_metrics.is_empty() {
-    update_gpu_histories(resources, &gpu_metrics);
-  }
-
-  gpu_metrics
-}
-
-#[cfg(target_os = "macos")]
-pub async fn sample_gpu(resources: &MonitorResources) -> Vec<GpuSample> {
-  use hwviz_core::infrastructure::providers::macos::io_kit::iokit_info;
-  use hwviz_core::infrastructure::providers::macos::{gpu, gpu_info};
-
-  static CACHED_GPU_NAME: tokio::sync::OnceCell<String> =
-    tokio::sync::OnceCell::const_new();
-
-  // Start the IOKit usage sampler thread (first call only)
-  let _ = gpu::init_gpu_usage_sampler_thread();
-
-  // Cache GPU name (fetched via IOKit + system_profiler on first call)
-  let gpu_name = CACHED_GPU_NAME
-    .get_or_init(|| async {
-      gpu_info::get_gpu_info()
-        .await
-        .ok()
-        .and_then(|list| list.into_iter().next())
-        .map(|info| info.name)
-        .unwrap_or_else(|| "Apple GPU".to_string())
-    })
-    .await;
-
-  // GPU usage (cached IOKit IOReport value, 0.0-1.0 → 0-100%)
-  let usage = gpu::read_gpu_usage_cached().map(|v| v * 100.0);
-
-  // Temperature: not available on macOS
-  let temperature: Option<f32> = None;
-
-  // Memory usage: convert IOKit bytes → KB (same unit as Windows NVAPI)
-  let memory_kb: Option<f32> =
-    tauri::async_runtime::spawn_blocking(iokit_info::get_gpu_memory_usage_from_iokit)
-      .await
-      .ok()
-      .flatten()
-      .and_then(|mem| mem.in_use_bytes)
-      .map(|bytes| (bytes / 1024) as f32);
-
-  let gpu_metrics = vec![GpuSample {
-    gpu_id: format!("iokit:{}", gpu_name),
-    name: gpu_name.clone(),
-    usage,
-    temperature,
-    dedicated_memory_kb: memory_kb,
-    cooler_level: None,
-    source: "IOKit".to_string(),
-  }];
-  update_gpu_histories(resources, &gpu_metrics);
-
-  gpu_metrics
-}
-
-/// Collect GPU metrics on Linux using the existing platform layer
-/// (DRM/sysfs for AMD, DRM for Intel).
-#[cfg(target_os = "linux")]
-async fn collect_linux_gpu_metrics() -> Vec<GpuSample> {
-  use hwviz_core::infrastructure::providers::drm_sys;
-  use hwviz_core::infrastructure::providers::hwmon;
-  use hwviz_core::infrastructure::providers::lspci;
-
-  let mut metrics: Vec<GpuSample> = Vec::new();
-  let card_ids = drm_sys::get_all_card_ids();
-
-  // Lazily fetched on first AMD card (avoids forking lspci on Intel-only systems
-  // while still fetching only once when multiple AMD cards are present)
-  let mut lspci_output: Option<Option<String>> = None;
-
-  for card_id in card_ids {
-    let vendor = drm_sys::detect_gpu_vendor(card_id);
-
-    let (name, usage, temperature, source) = match vendor {
-      drm_sys::GpuVendor::Amd => {
-        let cached = lspci_output.get_or_insert_with(lspci::get_lspci_nn_output);
-        let name = drm_sys::get_card_bdf(card_id)
-          .and_then(|bdf| {
-            cached
-              .as_deref()
-              .and_then(|out| lspci::parse_gpu_name_by_bdf(out, &bdf))
-          })
-          .unwrap_or_else(|| format!("AMD GPU (card{})", card_id));
-        let usage = drm_sys::get_amd_gpu_usage(card_id as u32)
-          .await
-          .map(|u| (u * 100.0) as f32)
-          .ok();
-        let temperature = hwmon::read_hwmon_temperatures(card_id)
-          .ok()
-          .and_then(|temps| temps.first().map(|t| t.value as f32));
-        (name, usage, temperature, "DRM (AMD)".to_string())
-      }
-      drm_sys::GpuVendor::Intel => {
-        let usage = drm_sys::get_intel_gpu_usage()
-          .await
-          .map(|u| (u * 100.0) as f32)
-          .ok();
-        (
-          format!("Intel GPU (card{})", card_id),
-          usage,
-          None,
-          "DRM (Intel)".to_string(),
-        )
-      }
-      _ => continue,
-    };
-
-    let gpu_id = drm_sys::get_card_bdf(card_id)
-      .map(|bdf| format!("pci:{bdf}"))
-      .unwrap_or_else(|| format!("drm:card{card_id}"));
-
-    // VRAM usage is not available on Linux
-    metrics.push(GpuSample {
-      gpu_id,
-      name,
-      usage,
-      temperature,
-      dedicated_memory_kb: None,
-      cooler_level: None,
-      source,
-    });
-  }
-
-  metrics
-}
-
-fn update_gpu_histories(resources: &MonitorResources, gpu_metrics: &[GpuSample]) {
-  let mut usage_histories = resources.gpu_usage_histories.lock().unwrap();
-  let mut temp_histories = resources.gpu_temperature_histories.lock().unwrap();
-  let mut mem_histories = resources.gpu_dedicated_memory_histories.lock().unwrap();
-  let mut name_map = resources.gpu_name_map.lock().unwrap();
-
-  gpu_metrics.iter().for_each(|sample| {
-    name_map.insert(sample.gpu_id.clone(), sample.name.clone());
-
-    if let Some(usage) = sample.usage {
-      let usage_history = usage_histories.entry(sample.gpu_id.clone()).or_default();
-      if usage_history.len() >= HARDWARE_HISTORY_BUFFER_SIZE {
-        usage_history.pop_front();
-      }
-      usage_history.push_back(usage);
-    }
-    if let Some(temperature) = sample.temperature {
-      let temp_history = temp_histories.entry(sample.gpu_id.clone()).or_default();
-      if temp_history.len() >= HARDWARE_HISTORY_BUFFER_SIZE {
-        temp_history.pop_front();
-      }
-      temp_history.push_back(temperature as i32);
-    }
-    if let Some(memory_usage) = sample.dedicated_memory_kb {
-      let mem_history = mem_histories.entry(sample.gpu_id.clone()).or_default();
-      if mem_history.len() >= HARDWARE_HISTORY_BUFFER_SIZE {
-        mem_history.pop_front();
-      }
-      mem_history.push_back(memory_usage as i32);
-    }
-  });
-}
-
-///
-/// ## CPU usage history
-///
-/// (Last `seconds` from newest, max MAX_HISTORY_QUERY_DURATION_SECONDS) collected in reverse order
-///
-pub fn cpu_usage_history(state: &HardwareMonitorState, seconds: u32) -> Vec<f32> {
-  let history = state.cpu_history.lock().unwrap();
-  let take_n = seconds.min(MAX_HISTORY_QUERY_DURATION_SECONDS) as usize;
-
-  history.iter().rev().take(take_n).cloned().collect()
-}
-
-///
-/// ## Memory usage history
-///
-/// (Last `seconds` from newest, max MAX_HISTORY_QUERY_DURATION_SECONDS) collected in reverse order
-///
-pub fn memory_usage_history(state: &HardwareMonitorState, seconds: u32) -> Vec<f32> {
-  let history = state.memory_history.lock().unwrap();
-  let take_n = seconds.min(MAX_HISTORY_QUERY_DURATION_SECONDS) as usize;
-
-  history.iter().rev().take(take_n).cloned().collect()
-}
-
-///
-/// ## GPU usage history
-///
-/// Returns the last `seconds` entries (newest first, max MAX_HISTORY_QUERY_DURATION_SECONDS)
-/// for the GPU identified by `gpu_id`.
-///
-pub fn gpu_usage_history(
-  state: &HardwareMonitorState,
-  gpu_id: &str,
-  seconds: u32,
-) -> Vec<f32> {
-  let histories = state.gpu_usage_histories.lock().unwrap();
-  let take_n = seconds.min(MAX_HISTORY_QUERY_DURATION_SECONDS) as usize;
-
-  histories
-    .get(gpu_id)
-    .map(|h| h.iter().rev().take(take_n).cloned().collect())
-    .unwrap_or_default()
-}
-
-fn push_history(history: &Arc<Mutex<VecDeque<f32>>>, value: f32) {
-  let mut h = history.lock().unwrap();
-  if h.len() >= HARDWARE_HISTORY_BUFFER_SIZE {
-    h.pop_front();
-  }
-  h.push_back(value);
-}
-
-#[cfg(test)]
-mod tests {
-  use super::*;
-  use std::collections::HashMap;
-
-  // ── Helpers ──
-
-  fn create_test_state(history_data: Vec<f32>) -> HardwareMonitorState {
-    let history: VecDeque<f32> = history_data.into();
-    HardwareMonitorState {
-      system: Arc::new(Mutex::new(sysinfo::System::new())),
-      cpu_history: Arc::new(Mutex::new(history.clone())),
-      memory_history: Arc::new(Mutex::new(history)),
-      process_cpu_histories: Arc::new(Mutex::new(HashMap::new())),
-      process_memory_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_usage_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_temperature_histories: Arc::new(Mutex::new(HashMap::new())),
-    }
-  }
-
-  fn create_test_state_with_gpu_history(
-    gpu_id: &str,
-    gpu_data: Vec<f32>,
-  ) -> HardwareMonitorState {
-    let mut gpu_histories = HashMap::new();
-    gpu_histories.insert(gpu_id.to_string(), VecDeque::from(gpu_data));
-    HardwareMonitorState {
-      system: Arc::new(Mutex::new(sysinfo::System::new())),
-      cpu_history: Arc::new(Mutex::new(VecDeque::new())),
-      memory_history: Arc::new(Mutex::new(VecDeque::new())),
-      process_cpu_histories: Arc::new(Mutex::new(HashMap::new())),
-      process_memory_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_usage_histories: Arc::new(Mutex::new(gpu_histories)),
-      gpu_temperature_histories: Arc::new(Mutex::new(HashMap::new())),
-    }
-  }
-
-  fn create_test_resources() -> MonitorResources {
-    MonitorResources {
-      system: Arc::new(Mutex::new(sysinfo::System::new())),
-      cpu_history: Arc::new(Mutex::new(VecDeque::new())),
-      memory_history: Arc::new(Mutex::new(VecDeque::new())),
-      process_cpu_histories: Arc::new(Mutex::new(HashMap::new())),
-      process_memory_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_usage_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_temperature_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_dedicated_memory_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_name_map: Arc::new(Mutex::new(HashMap::new())),
-    }
-  }
-
-  // ── calculate_memory_usage_percentage ──
-
-  #[test]
-  fn memory_usage_percentage_total_zero() {
-    assert_eq!(calculate_memory_usage_percentage(1000, 0), 0.0);
-  }
-
-  #[test]
-  fn memory_usage_percentage_half_used() {
-    assert_eq!(calculate_memory_usage_percentage(500, 1000), 50.0);
-  }
-
-  #[test]
-  fn memory_usage_percentage_fully_used() {
-    assert_eq!(calculate_memory_usage_percentage(1000, 1000), 100.0);
-  }
-
-  #[test]
-  fn memory_usage_percentage_zero_used() {
-    assert_eq!(calculate_memory_usage_percentage(0, 1000), 0.0);
-  }
-
-  #[test]
-  fn memory_usage_percentage_rounding() {
-    assert_eq!(calculate_memory_usage_percentage(333, 1000), 33.0);
-  }
-
-  #[test]
-  fn memory_usage_percentage_large_values() {
-    assert_eq!(
-      calculate_memory_usage_percentage(16_000_000_000, 32_000_000_000),
-      50.0
-    );
-  }
-
-  // ── push_history ──
-
-  #[test]
-  fn push_history_to_empty() {
-    let history = Arc::new(Mutex::new(VecDeque::new()));
-    push_history(&history, 1.0);
-    let h = history.lock().unwrap();
-    assert_eq!(h.len(), 1);
-    assert_eq!(h[0], 1.0);
-  }
-
-  #[test]
-  fn push_history_under_capacity() {
-    let history = Arc::new(Mutex::new(VecDeque::from(vec![0.0; 10])));
-    push_history(&history, 1.0);
-    assert_eq!(history.lock().unwrap().len(), 11);
-  }
-
-  #[test]
-  fn push_history_at_capacity_evicts_oldest() {
-    let initial: VecDeque<f32> = (0..HARDWARE_HISTORY_BUFFER_SIZE)
-      .map(|i| i as f32)
-      .collect();
-    let history = Arc::new(Mutex::new(initial));
-    push_history(&history, 999.0);
-
-    let h = history.lock().unwrap();
-    assert_eq!(h.len(), HARDWARE_HISTORY_BUFFER_SIZE);
-    assert_eq!(*h.front().unwrap(), 1.0); // 0.0 was evicted
-    assert_eq!(*h.back().unwrap(), 999.0);
-  }
-
-  #[test]
-  fn push_history_preserves_order() {
-    let history = Arc::new(Mutex::new(VecDeque::new()));
-    push_history(&history, 1.0);
-    push_history(&history, 2.0);
-    push_history(&history, 3.0);
-    let h = history.lock().unwrap();
-    assert_eq!(h.iter().cloned().collect::<Vec<_>>(), vec![1.0, 2.0, 3.0]);
-  }
-
-  // ── cpu_usage_history ──
-
-  #[test]
-  fn cpu_usage_history_returns_last_n_reversed() {
-    let state = create_test_state(vec![1.0, 2.0, 3.0, 4.0, 5.0]);
-    let result = cpu_usage_history(&state, 3);
-    assert_eq!(result, vec![5.0, 4.0, 3.0]);
-  }
-
-  #[test]
-  fn cpu_usage_history_clamped_to_available() {
-    let state = create_test_state(vec![1.0, 2.0]);
-    let result = cpu_usage_history(&state, 10);
-    assert_eq!(result, vec![2.0, 1.0]);
-  }
-
-  #[test]
-  fn cpu_usage_history_clamped_to_max_duration() {
-    let data: Vec<f32> = (0..5000).map(|i| i as f32).collect();
-    let state = create_test_state(data);
-    let result = cpu_usage_history(&state, u32::MAX);
-    assert_eq!(result.len(), MAX_HISTORY_QUERY_DURATION_SECONDS as usize);
-  }
-
-  #[test]
-  fn cpu_usage_history_empty() {
-    let state = create_test_state(vec![]);
-    let result = cpu_usage_history(&state, 5);
-    assert!(result.is_empty());
-  }
-
-  // ── memory_usage_history ──
-
-  #[test]
-  fn memory_usage_history_returns_last_n_reversed() {
-    let state = create_test_state(vec![10.0, 20.0, 30.0]);
-    let result = memory_usage_history(&state, 2);
-    assert_eq!(result, vec![30.0, 20.0]);
-  }
-
-  #[test]
-  fn memory_usage_history_empty() {
-    let state = create_test_state(vec![]);
-    let result = memory_usage_history(&state, 5);
-    assert!(result.is_empty());
-  }
-
-  // ── gpu_usage_history ──
-
-  #[test]
-  fn gpu_usage_history_returns_last_n_reversed() {
-    let state = create_test_state_with_gpu_history("gpu:0", vec![40.0, 50.0, 60.0, 70.0]);
-    let result = gpu_usage_history(&state, "gpu:0", 3);
-    assert_eq!(result, vec![70.0, 60.0, 50.0]);
-  }
-
-  #[test]
-  fn gpu_usage_history_empty() {
-    let state = create_test_state(vec![]);
-    let result = gpu_usage_history(&state, "gpu:0", 5);
-    assert!(result.is_empty());
-  }
-
-  #[test]
-  fn gpu_usage_history_unknown_gpu_id_returns_empty() {
-    let state = create_test_state_with_gpu_history("gpu:0", vec![10.0, 20.0]);
-    let result = gpu_usage_history(&state, "gpu:nonexistent", 5);
-    assert!(result.is_empty());
-  }
-
-  #[test]
-  fn gpu_usage_history_by_gpu_id() {
-    let mut gpu_histories = HashMap::new();
-    gpu_histories.insert("gpu:0".to_string(), VecDeque::from(vec![10.0, 20.0]));
-    gpu_histories.insert("gpu:1".to_string(), VecDeque::from(vec![90.0, 80.0]));
-    let state = HardwareMonitorState {
-      system: Arc::new(Mutex::new(sysinfo::System::new())),
-      cpu_history: Arc::new(Mutex::new(VecDeque::new())),
-      memory_history: Arc::new(Mutex::new(VecDeque::new())),
-      process_cpu_histories: Arc::new(Mutex::new(HashMap::new())),
-      process_memory_histories: Arc::new(Mutex::new(HashMap::new())),
-      gpu_usage_histories: Arc::new(Mutex::new(gpu_histories)),
-      gpu_temperature_histories: Arc::new(Mutex::new(HashMap::new())),
-    };
-    assert_eq!(gpu_usage_history(&state, "gpu:0", 5), vec![20.0, 10.0]);
-    assert_eq!(gpu_usage_history(&state, "gpu:1", 5), vec![80.0, 90.0]);
-  }
-
-  #[test]
-  fn gpu_usage_history_clamped_to_max_duration() {
-    let data: Vec<f32> = (0..5000).map(|i| i as f32).collect();
-    let state = create_test_state_with_gpu_history("gpu:0", data);
-    let result = gpu_usage_history(&state, "gpu:0", u32::MAX);
-    assert_eq!(result.len(), MAX_HISTORY_QUERY_DURATION_SECONDS as usize);
-  }
-
-  // ── update_gpu_histories ──
-
-  fn make_gpu_sample(
-    gpu_id: &str,
-    name: &str,
-    usage: Option<f32>,
-    temperature: Option<f32>,
-    dedicated_memory_kb: Option<f32>,
-  ) -> GpuSample {
-    GpuSample {
-      gpu_id: gpu_id.to_string(),
-      name: name.to_string(),
-      usage,
-      temperature,
-      dedicated_memory_kb,
-      cooler_level: None,
-      source: "Test".to_string(),
-    }
-  }
-
-  #[test]
-  fn gpu_sample_has_gpu_id_field() {
-    let sample = make_gpu_sample("pci:0:2.0", "TestGPU", Some(50.0), None, None);
-    assert_eq!(sample.gpu_id, "pci:0:2.0");
-    assert_eq!(sample.name, "TestGPU");
-  }
-
-  #[test]
-  fn update_gpu_histories_adds_new_gpu() {
-    let resources = create_test_resources();
-    let samples = vec![make_gpu_sample(
-      "gpu:0",
-      "TestGPU",
-      Some(50.0),
-      Some(70.0),
-      Some(1024.0),
-    )];
-    update_gpu_histories(&resources, &samples);
-
-    let usage = resources.gpu_usage_histories.lock().unwrap();
-    assert_eq!(usage.get("gpu:0").unwrap().len(), 1);
-    assert_eq!(*usage.get("gpu:0").unwrap().back().unwrap(), 50.0);
-
-    let temp = resources.gpu_temperature_histories.lock().unwrap();
-    assert_eq!(*temp.get("gpu:0").unwrap().back().unwrap(), 70);
-
-    let mem = resources.gpu_dedicated_memory_histories.lock().unwrap();
-    assert_eq!(*mem.get("gpu:0").unwrap().back().unwrap(), 1024);
-  }
-
-  #[test]
-  fn update_gpu_histories_none_metrics_skipped() {
-    let resources = create_test_resources();
-    let samples = vec![make_gpu_sample("gpu:0", "NoData", None, None, None)];
-    update_gpu_histories(&resources, &samples);
-
-    assert!(resources.gpu_usage_histories.lock().unwrap().is_empty());
-    assert!(
-      resources
-        .gpu_temperature_histories
-        .lock()
-        .unwrap()
-        .is_empty()
-    );
-    assert!(
-      resources
-        .gpu_dedicated_memory_histories
-        .lock()
-        .unwrap()
-        .is_empty()
-    );
-  }
-
-  #[test]
-  fn update_gpu_histories_partial_metrics() {
-    let resources = create_test_resources();
-    let samples = vec![make_gpu_sample("gpu:0", "Partial", Some(80.0), None, None)];
-    update_gpu_histories(&resources, &samples);
-
-    assert_eq!(resources.gpu_usage_histories.lock().unwrap().len(), 1);
-    assert!(
-      resources
-        .gpu_temperature_histories
-        .lock()
-        .unwrap()
-        .is_empty()
-    );
-    assert!(
-      resources
-        .gpu_dedicated_memory_histories
-        .lock()
-        .unwrap()
-        .is_empty()
-    );
-  }
-
-  #[test]
-  fn update_gpu_histories_multiple_gpus() {
-    let resources = create_test_resources();
-    let samples = vec![
-      make_gpu_sample("pci:0:2.0", "GPU_A", Some(10.0), None, None),
-      make_gpu_sample("pci:0:3.0", "GPU_B", Some(90.0), None, None),
-    ];
-    update_gpu_histories(&resources, &samples);
-
-    let usage = resources.gpu_usage_histories.lock().unwrap();
-    assert_eq!(usage.len(), 2);
-    assert!(usage.contains_key("pci:0:2.0"));
-    assert!(usage.contains_key("pci:0:3.0"));
-  }
-
-  #[test]
-  fn update_gpu_histories_same_name_different_id() {
-    let resources = create_test_resources();
-    let samples = vec![
-      make_gpu_sample("pci:0:2.0", "NVIDIA RTX 4090", Some(30.0), None, None),
-      make_gpu_sample("pci:0:3.0", "NVIDIA RTX 4090", Some(70.0), None, None),
-    ];
-    update_gpu_histories(&resources, &samples);
-
-    let usage = resources.gpu_usage_histories.lock().unwrap();
-    assert_eq!(usage.len(), 2);
-    assert_eq!(*usage.get("pci:0:2.0").unwrap().back().unwrap(), 30.0);
-    assert_eq!(*usage.get("pci:0:3.0").unwrap().back().unwrap(), 70.0);
-  }
-
-  #[test]
-  fn update_gpu_histories_respects_buffer_cap() {
-    let resources = create_test_resources();
-    // Fill to capacity
-    for i in 0..HARDWARE_HISTORY_BUFFER_SIZE {
-      let samples = vec![make_gpu_sample("gpu:0", "GPU", Some(i as f32), None, None)];
-      update_gpu_histories(&resources, &samples);
-    }
-    // Add one more
-    let samples = vec![make_gpu_sample("gpu:0", "GPU", Some(999.0), None, None)];
-    update_gpu_histories(&resources, &samples);
-
-    let usage = resources.gpu_usage_histories.lock().unwrap();
-    let history = usage.get("gpu:0").unwrap();
-    assert_eq!(history.len(), HARDWARE_HISTORY_BUFFER_SIZE);
-    assert_eq!(*history.back().unwrap(), 999.0);
-    assert_eq!(*history.front().unwrap(), 1.0); // 0.0 was evicted
-  }
-
-  #[test]
-  fn update_gpu_histories_populates_name_map() {
-    let resources = create_test_resources();
-    let samples = vec![
-      make_gpu_sample("pci:0:2.0", "RTX 4090", Some(50.0), None, None),
-      make_gpu_sample("pci:0:3.0", "RX 7900 XTX", Some(80.0), None, None),
-    ];
-    update_gpu_histories(&resources, &samples);
-
-    let name_map = resources.gpu_name_map.lock().unwrap();
-    assert_eq!(name_map.get("pci:0:2.0").unwrap(), "RTX 4090");
-    assert_eq!(name_map.get("pci:0:3.0").unwrap(), "RX 7900 XTX");
-  }
-
-  // ── update_process_histories ──
-
-  #[test]
-  fn update_process_histories_new_process() {
-    let resources = create_test_resources();
-    let pid = sysinfo::Pid::from_u32(100);
-    let metrics = vec![(pid, 50.0_f32, 256.0_f32)];
-    update_process_histories(&resources, &metrics);
-
-    let cpu_h = resources.process_cpu_histories.lock().unwrap();
-    assert_eq!(cpu_h.get(&pid).unwrap().len(), 1);
-    assert_eq!(*cpu_h.get(&pid).unwrap().back().unwrap(), 50.0);
-
-    let mem_h = resources.process_memory_histories.lock().unwrap();
-    assert_eq!(*mem_h.get(&pid).unwrap().back().unwrap(), 256.0);
-  }
-
-  #[test]
-  fn update_process_histories_existing_process() {
-    let resources = create_test_resources();
-    let pid = sysinfo::Pid::from_u32(200);
-    let metrics = vec![(pid, 10.0_f32, 128.0_f32)];
-    update_process_histories(&resources, &metrics);
-    let metrics2 = vec![(pid, 20.0_f32, 256.0_f32)];
-    update_process_histories(&resources, &metrics2);
-
-    let cpu_h = resources.process_cpu_histories.lock().unwrap();
-    assert_eq!(cpu_h.get(&pid).unwrap().len(), 2);
-  }
-
-  #[test]
-  fn update_process_histories_respects_buffer_cap() {
-    let resources = create_test_resources();
-    let pid = sysinfo::Pid::from_u32(300);
-    for i in 0..HARDWARE_HISTORY_BUFFER_SIZE + 5 {
-      let metrics = vec![(pid, i as f32, i as f32)];
-      update_process_histories(&resources, &metrics);
-    }
-
-    let cpu_h = resources.process_cpu_histories.lock().unwrap();
-    assert_eq!(cpu_h.get(&pid).unwrap().len(), HARDWARE_HISTORY_BUFFER_SIZE);
-  }
-
-  // ── resolve_gpu_name_from_map ──
-
-  // ── update_gpu_histories: multi-AMD GPU name collision regression ──
-
-  #[test]
-  fn update_gpu_histories_two_amd_gpus_separate_entries() {
-    let resources = create_test_resources();
-    let samples = vec![
-      GpuSample {
-        gpu_id: "pci:06:00.0".to_string(),
-        name: "06:00.0 VGA compatible controller [0300]: Advanced Micro Devices, Inc. [AMD/ATI] Navi 31 [Radeon RX 7900 XTX] [1002:744c]".to_string(),
-        usage: Some(80.0),
-        temperature: Some(65.0),
-        dedicated_memory_kb: None,
-        cooler_level: None,
-        source: "DRM (AMD)".to_string(),
-      },
-      GpuSample {
-        gpu_id: "pci:0a:00.0".to_string(),
-        name: "0a:00.0 VGA compatible controller [0300]: Advanced Micro Devices, Inc. [AMD/ATI] Renoir [Radeon Vega Series] [1002:1636]".to_string(),
-        usage: Some(15.0),
-        temperature: Some(45.0),
-        dedicated_memory_kb: None,
-        cooler_level: None,
-        source: "DRM (AMD)".to_string(),
-      },
-    ];
-    update_gpu_histories(&resources, &samples);
-
-    let usage = resources.gpu_usage_histories.lock().unwrap();
-    assert_eq!(
-      usage.len(),
-      2,
-      "Two AMD GPUs must have separate history entries"
-    );
-
-    let temp = resources.gpu_temperature_histories.lock().unwrap();
-    assert_eq!(temp.len(), 2);
-  }
-
-  // ── resolve_gpu_name_from_map ──
-
-  #[test]
-  fn resolve_gpu_name_returns_dxgi_name_on_bdf_hit() {
-    let mut map = HashMap::new();
-    map.insert((3, 0, 0), "AMD Radeon RX 7900 XTX".to_string());
-    let result = resolve_gpu_name_from_map(&map, "Radeon RX 7900 XTX", 3, 0, 0);
-    assert_eq!(result, "AMD Radeon RX 7900 XTX");
-  }
-
-  #[test]
-  fn resolve_gpu_name_falls_back_to_adl_name_on_miss() {
-    let map = HashMap::new(); // empty map
-    let result = resolve_gpu_name_from_map(&map, "Radeon RX 7900 XTX", 3, 0, 0);
-    assert_eq!(result, "Radeon RX 7900 XTX");
-  }
-
-  #[test]
-  fn resolve_gpu_name_distinguishes_by_bdf() {
-    let mut map = HashMap::new();
-    map.insert((3, 0, 0), "GPU on bus 3".to_string());
-    map.insert((6, 0, 0), "GPU on bus 6".to_string());
-
-    assert_eq!(
-      resolve_gpu_name_from_map(&map, "fallback", 3, 0, 0),
-      "GPU on bus 3"
-    );
-    assert_eq!(
-      resolve_gpu_name_from_map(&map, "fallback", 6, 0, 0),
-      "GPU on bus 6"
-    );
-    // Different BDF not in map → fallback
-    assert_eq!(
-      resolve_gpu_name_from_map(&map, "fallback", 9, 0, 0),
-      "fallback"
-    );
-  }
-
-  #[test]
-  fn resolve_gpu_name_handles_duplicate_bdf() {
-    let mut map = HashMap::new();
-    // Last insert wins (HashMap behavior)
-    map.insert((3, 0, 0), "First".to_string());
-    map.insert((3, 0, 0), "Second".to_string());
-    assert_eq!(
-      resolve_gpu_name_from_map(&map, "fallback", 3, 0, 0),
-      "Second"
-    );
-  }
-}
